# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    [OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#management)

    HTTP API to be implemented by Event Transmitters. This API can be used by
    Event Receivers to query and update the Event Stream configuration and
    status, to add and remove subjects, and to trigger verification.
  license:
    name: MIT
  title: Stream Management API for OpenID Shared Security Events
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /stream:
    delete:
      description: |-
        An Event Receiver removes the configuration of a stream by making an HTTP DELETE request to the Configuration
        Endpoint. On receiving a request the Event Transmitter responds with a 200 OK response if the configuration was
        successfully removed.
      responses:
        "200":
          description: |-
            On receiving a request the Event Transmitter responds with a 200 OK response if the configuration was
            successfully removed.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Authorization failed or is missing
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The Event Receiver is not allowed to update the stream configuration.
      security:
      - BearerAuth: []
      summary: Request to remove the configuration of an event stream
      tags:
      - StreamManagement
      x-accepts: application/json
      x-tags:
      - tag: StreamManagement
    get:
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#reading-a-streams-configuration)

        An Event Receiver gets the current configuration of a stream by making an HTTP GET request to the Configuration
        Endpoint. On receiving a valid request the Event Transmitter responds with a 200 OK response containing a JSON
        representation of the stream’s configuration in the body.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamConfiguration'
          description: The Event Stream's configuration
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Request body cannot be parsed or the request is otherwise invalid
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The Event Receiver is not allowed to read the stream configuration.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: There is no Event Stream configured for this Event Receiver.
      security:
      - BearerAuth: []
      summary: Request to retrieve the configuration of an event stream
      tags:
      - StreamManagement
      x-accepts: application/json
      x-tags:
      - tag: StreamManagement
    post:
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#updating-a-streams-configuration)

        An Event Receiver updates the current configuration of a stream by making an HTTP POST request to the Configuration
        Endpoint. The POST body contains a JSON representation of the updated configuration. On receiving a valid request
        the Event Transmitter responds with a 200 OK response containing a JSON representation of the updated stream
        configuration in the body.

        The full set of editable properties must be present in the POST body, not only the ones that are specifically
        intended to be changed. Missing properties SHOULD be interpreted as requested to be deleted. Event Receivers should
        read the configuration first, modify the JSON representation, then make an update request.

        Properties that cannot be updated MAY be present, but they MUST match the expected value.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamConfiguration'
        description: Request parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamConfiguration'
          description: The Event Stream's updated configuration
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: |-
            The update request has been accepted, but not processed.
            Receiver MAY try the same request later in order to get processing result.

            If a Receiver makes a request to update the stream to enable it for a specific Subject, and the Transmitter is
            unable to decide whether or not to complete the request, then the Transmitter MUST respond with a 202 status
            code.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Request body cannot be parsed or the request is otherwise invalid
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Authorization failed or is missing
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The Event Receiver is not allowed to update the stream configuration.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: There is no Event Stream configured for this Event Receiver.
      security:
      - BearerAuth: []
      summary: Request to update the configuration of an event stream
      tags:
      - StreamManagement
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: StreamManagement
...
components:
...
  schemas:
...
    StreamConfiguration:
      description: |
        JSON Object describing and Event Stream's configuration
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#stream-config)"
      example:
        aud: https://popular-app.com
        delivery:
          method: https://schemas.openid.net/secevent/risc/delivery-method/poll
          endpoint_url: null
        events_supported:
        - https://schemas.openid.net/secevent/caep/event-type/session-revoked
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        iss: https://most-secure.com
        format: format
        events_delivered:
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        events_requested:
        - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
        min_verification_interval: 0
      properties:
        iss:
          description: |-
            Read-Only.
            A URL using the https scheme with no query or fragment component that the Transmitter asserts as its Issuer
            Identifier. This MUST be identical to the iss Claim value in Security Event Tokens issued from this Transmitter.
          example: https://most-secure.com
          format: uri
          type: string
        aud:
          $ref: '#/components/schemas/StreamConfiguration_aud'
        events_supported:
          description: |-
            Read-Only.
            An array of URIs identifying the set of events supported by the Transmitter for this Receiver.
            If omitted, Event Transmitters SHOULD make this set available to the Event Receiver via some other means
            (e.g. publishing it in online documentation).
          example:
          - https://schemas.openid.net/secevent/caep/event-type/session-revoked
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            format: uri
            type: string
          type: array
        events_requested:
          description: |-
            Read-Write.
            An array of URIs identifying the set of events that the Receiver requested.
            A Receiver SHOULD request only the events that it understands and it can act on.
            This is configurable by the Receiver.
          example:
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            format: uri
            type: string
          type: array
        events_delivered:
          description: |-
            Read-Only.
            An array of URIs which is the intersection of events_supported and events_requested.
            These events MAY be delivered over the Event Stream.
          example:
          - https://schemas.openid.net/secevent/risc/event-type/credential-compromise
          items:
            format: uri
            type: string
          type: array
        delivery:
          $ref: '#/components/schemas/StreamConfiguration_delivery'
        min_verification_interval:
          description: |-
            Read-Only.
            An integer indicating the minimum amount of time in seconds that must pass in between verification requests.
            If an Event Receiver submits verification requests more frequently than this, the Event Transmitter MAY respond
            with a 429 status code. An Event Transmitter SHOULD NOT respond with a 429 status code if an Event Receiver is not
            exceeding this frequency.
          type: integer
        format:
          description: |-
            Read-Write.
            The Subject Identifier Format that the Receiver wants for the events.
            If not set then the Transmitter might decide to use a type that discloses more information than necessary.
          type: string
      required:
      - delivery
      - events_requested
      title: Event Stream Configuration
      type: object
...
    PollDeliveryMethod:
      example:
        method: https://schemas.openid.net/secevent/risc/delivery-method/poll
        endpoint_url: https://transmitter.most-secure.com/poll
      properties:
        method:
          default: https://schemas.openid.net/secevent/risc/delivery-method/poll
          type: string
        endpoint_url:
          description: The URL where events can be retrieved from. This is specified
            by the Transmitter.
          format: uri
          nullable: true
          type: string
      required:
      - method
      type: object
    PushDeliveryMethod:
      example:
        method: https://schemas.openid.net/secevent/risc/delivery-method/push
        endpoint_url: https://receiver.example.com/push
      properties:
        method:
          default: https://schemas.openid.net/secevent/risc/delivery-method/push
          type: string
        endpoint_url:
          description: The URL where events are pushed through HTTP POST. This is
            set by the Receiver.
          format: uri
          type: string
        authorization_header:
          description: |-
            The HTTP Authorization header that the Transmitter MUST set with each event delivery,
            if the configuration is present. The value is optional and it is set by the Receiver.
          type: string
      required:
      - endpoint_url
      - method
      type: object
...
    Error:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
    StreamConfiguration_aud:
      description: |-
        Read-Only.
        A string or an array of strings containing an audience claim as defined in
        [JSON Web Token (JWT)](https://openid.net/specs/openid-sse-framework-1_0.html#RFC7519) that identifies
        the Event Receiver(s) for the Event Stream. This property cannot be updated. If multiple Receivers are specified
        then the Transmitter SHOULD know that these Receivers are the same entity.
      example: https://popular-app.com
      oneOf:
      - format: uri
        type: string
      - items:
          format: uri
          type: string
        type: array
      x-one-of-name: StreamConfigurationAud
    StreamConfiguration_delivery:
      description: |-
        Read-Write.
        A JSON object containing a set of name/value pairs specifying configuration parameters for the SET delivery
        method. The actual delivery method is identified by the special key method with the value being a URI as defined
        in [Section 11.2.1](https://openid.net/specs/openid-sse-framework-1_0.html#delivery-meta).
      discriminator:
        propertyName: method
      example:
        method: https://schemas.openid.net/secevent/risc/delivery-method/poll
        endpoint_url: null
      oneOf:
      - $ref: '#/components/schemas/PushDeliveryMethod'
      - $ref: '#/components/schemas/PollDeliveryMethod'
      x-one-of-name: StreamConfigurationDelivery
...
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http